refactor : data files refactored. [note tests module should later get out of the installed modules]
